openrisc: Return ms granularity with <1000 Hz timer irq period
Since CONFIG_HZ is mandatory to be set to 1000 Hz,
a CONFIG_SYS_OPENRISC_TMR_HZ knob is introduced to be
able to tweak the period of the timer exception.

Setting this value to <1000 would hurt the granularity of
get_timer, so the current value of the timer has to be added to
the accumulated value from full timer periods.